requirement.txt
pytorch
torchvision
timm
mxnet
opencv-python